Fixed some parts of ffmpeg convert to flac, I hadn't tested this, but -aq / -q:a do not actually do anything, ffmpeg reverts to default quality 5 always. This is fixed by using -compression_level instead, which works as expected. Note that for some reason, -compression_level accepts 0-12 instead of the standard flac 0-8, and in fact, 9-12 do make small changes in file size. However, given how small the changes are, I'm leaving max flac quality as 8 since that is what most people know about.
I also added a silly feature to allow acxi to convert mp3 to flac,
which is not recommended, not particularly smart, but acxi will now
do it, since I found at least one case where that is worth doing.
Note that I further found another ffmpeg oddity there, when doing
a conversion to flac from mp3, ffmpeg defaulted to 24 bit flac output,
which is obviously silly, so I added -sample_fsmt s16 for mp3 input
type there, which forces to 16 bit flac.

I wish I'd tested the quality levels a bit more for shn-> flac conversion,
but since the default of 5 is similar to the acxi flac default of 4, I never
really noticed the face that -aq was not working.