Big update, removed -Z/--zero, added more flexible --start [0-xx] which allows you to select any track number to start on. This lets you do 0, which also will modify the track totals so the expected track y/xx track number/total tracks will show in a more intuitive way. It also lets you handle releases where for some reason the first x tracks have been removed, in that case, it will show the track total as it would have been had the missing tracks not been removed, and the tagged track numbers will match the actual file track numbers.
Second is a simarly useful in corner cases, --recurse [-1 - xx], which
allows you to limit the returned results to only recursion level you
specify. That lts you for example ignore a sub directory that may have
music or other files you don't processing to be done on. Be careful
with that one, use the --test option to see what would have happened
to get a feel for what these will do.